Delete ROX_SYSTEM_HEALTH_PF feature flag
