+++ ui/apps/platform/src/Containers/SystemHealth/Components/DiagnosticBundleForm.tsx
