+++ ui/apps/platform/src/Containers/SystemHealth/Components/GenerateDiagnosticBundle.tsx
