+++ ui/apps/platform/src/Containers/SystemHealth/utils/diagnosticBundleUtils.test.ts
