+++ ui/apps/platform/src/Containers/SystemHealth/utils/diagnosticBundleUtils.ts
