+++ ui/apps/platform/src/types/featureFlag.ts
@@ -9,7 +9,6 @@ export type FeatureFlagEnvVar =
     | 'ROX_NETWORK_GRAPH_PATTERNFLY'
     | 'ROX_POSTGRES_DATASTORE'
     | 'ROX_SYSLOG_EXTRA_FIELDS'
-    | 'ROX_SYSTEM_HEALTH_PF'
     | 'ROX_VULN_MGMT_REPORTING_ENHANCEMENTS'
     | 'ROX_VULN_MGMT_WORKLOAD_CVES'
     ;
